Move user model to models package
